requirements.txt
addict==2.4.0
annotated-doc==0.0.4
anyio==4.12.1
braceexpand==0.1.7
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.3.1
clip-benchmark==1.6.2
contourpy==1.3.2
cycler==0.12.1
Cython==3.2.4
exceptiongroup==1.3.1
filelock==3.20.0
fonttools==4.62.1
fsspec==2025.12.0
ftfy==6.3.1
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.4
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.5.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.8
mdurl==0.1.2
mmcv-full==1.7.2
-e git+https://github.com/open-mmlab/mmdetection.git@e9cae2d0787cd5c2fc6165a6061f92fa09e48fb1#egg=mmdet
mpmath==1.3.0
networkx==3.4.2
ninja==1.13.0
numpy==1.26.4
open_clip_torch==3.3.0
opencv-python-headless==4.13.0.92
packaging==26.0
pandas==2.3.3
pillow==12.0.0
platformdirs==4.9.4
pycocoevalcap==1.2
pycocotools==2.0.11
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
regex==2026.2.28
requests==2.28.1
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
sympy==1.14.0
terminaltables==3.1.10
threadpoolctl==3.6.0
timm==1.0.25
tokenizers==0.15.2
tomli==2.4.0
torch==2.1.1+cu121
torchaudio==2.1.1+cu121
torchvision==0.16.1+cu121
tqdm==4.67.3
transformers==4.36.1
triton==2.1.0
typer==0.24.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==1.26.13
wcwidth==0.6.0
webdataset==1.0.2
yapf==0.43.0
